pyproject.toml
[tool.poetry]
name = "task-management-system-server"
version = "0.1.0"
description = ""
authors = ["gaurav-gaikwad-12071998 <122888092+gaurav-gaikwad-12071998@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
asyncpg = "^0.30.0"
injector = "^0.22.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src:start_server"